Refactor to substantially reduce uncessary mutex usage and creation for scans. Big performance gains.

+++ squalr-cli/src/command_handlers/process/process_command.rs
@@ -22,7 +22,7 @@ pub enum ProcessCommand {
         #[structopt(short = "s", long)]
         include_system_processes: bool,
         #[structopt(short = "l", long)]
-        limit: Option<usize>,
+        limit: Option<u64>,
     },
     Close,
 }
